MAINT: refactor make_tree, load_tree
[CHANGED] moved functions into cogent3/core/tree.py to reduce
    complexity of cogent3/__init__.py
